feat(:angola:): start with utility function to format cvd codes
AFFECTS PACKAGES:
@esri/arcgis-rest-feature-service
