+++ packages/arcgis-rest-feature-service/src/index.ts
@@ -11,3 +11,4 @@ export * from "./updateAttachment";
 export * from "./deleteAttachments";
 export * from "./queryRelated";
 export * from "./getFeatureService";
+export * from "./formatCodedValues";
